fix: Rust/C# 编译通过 — 修复链接器、导入、项目配置
Rust (cargo build + test 5/5 passed):
- 用 rust-lld 替代 MSVC link.exe (环境缺 VS Build Tools)
- 设置 LIB 指向 Windows SDK 库路径
- 修复 Cargo.toml: 补充 errhandlingapi/winerror features, ctrlc 依赖
- 修复 pipe.rs: 补充 OsStrExt/GetLastError/winapi 导入, HANDLE Send 问题
- 清理 event.rs 未使用 import

C# (dotnet build 0/0 通过):
- 移除不存在的 NuGet 包 (System.IO.Pipes, Uwp.Notifications, H.NotifyIcon)
- 新增 AppSettings.cs (JSON 持久化) 替代 Properties.Settings.Default
- 新增 GlobalUsings.cs 解决 WPF/WinForms 类型歧义
- 关闭 PublishTrimmed (WPF 不支持)
- 恢复 UseWindowsForms (TrayService 依赖 NotifyIcon)
- 允许 unsafe 代码 (PixelMascot 指针操作)
- 创建 .sln 解决方案文件

+++ AgentScope.App/Controls/PixelMascot.cs
@@ -199,9 +199,9 @@ public class PixelMascot : FrameworkElement
                         var idx = ((y - offset) * SourceSize + x) * 4;
                         if (idx >= 0 && idx + 3 < frame.Length)
                         {
-                            frame[idx] = b;     // B
-                            frame[idx + 1] = g; // G
-                            frame[idx + 2] = r; // R
+                            frame[idx] = (byte)b;     // B
+                            frame[idx + 1] = (byte)g; // G
+                            frame[idx + 2] = (byte)r; // R
                             frame[idx + 3] = 255; // A
                         }
                     }
@@ -227,13 +227,13 @@ public class PixelMascot : FrameworkElement
         return frames;
     }
 
-    private static void DrawPixel(byte[] frame, int x, int y, byte r, byte g, byte b)
+    private static void DrawPixel(byte[] frame, int x, int y, int r, int g, int b)
     {
         if (x < 0 || x >= SourceSize || y < 0 || y >= SourceSize) return;
         var idx = (y * SourceSize + x) * 4;
-        frame[idx] = b;
-        frame[idx + 1] = g;
-        frame[idx + 2] = r;
+        frame[idx] = (byte)b;
+        frame[idx + 1] = (byte)g;
+        frame[idx + 2] = (byte)r;
         frame[idx + 3] = 255;
     }
 }
